update format of pinned notebooks for users that used the september release version
